ahc/web/requirements.txt
django-grappelli
django-simple-captcha
django-preferences
git+https://github.com/Govexec/ge-admin-login-captcha.git
